// src/model/createIndexRequest.ts
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.12.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { IndexConfig } from './indexConfig';

/**
* Request model for creating a new encrypted index.  Attributes:     index_config (Optional[Union[IndexIVFModel, IndexIVFPQModel, IndexIVFFlatModel]]):          Optional configuration model for the index.     index_key (str): A 32-byte encryption key as a hex string.     index_name (str): The name/identifier of the index.     embedding_model (Optional[str]): Optional embedding model name.     metric (Optional[str]): Optional distance metric.
*/
export class CreateIndexRequest {
    'indexConfig'?: IndexConfig | null;
    /**
    * 32-byte encryption key as hex string
    */
    'indexKey': string;
    /**
    * ID name
    */
    'indexName': string;
    'embeddingModel'?: string | null;
    'metric'?: string | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "indexConfig",
            "baseName": "index_config",
            "type": "IndexConfig"
        },
        {
            "name": "indexKey",
            "baseName": "index_key",
            "type": "string"
        },
        {
            "name": "indexName",
            "baseName": "index_name",
            "type": "string"
        },
        {
            "name": "embeddingModel",
            "baseName": "embedding_model",
            "type": "string"
        },
        {
            "name": "metric",
            "baseName": "metric",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return CreateIndexRequest.attributeTypeMap;
    }
}