// src/model/getResponseModel.ts
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.9.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { GetResultItemModel } from './getResultItemModel';

/**
* Response model for retrieving multiple encrypted index items.  Attributes:     results (List[GetResultItem]): A list of retrieved items with requested fields.
*/
export class GetResponseModel {
    'results': Array<GetResultItemModel>;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "results",
            "baseName": "results",
            "type": "Array<GetResultItemModel>"
        }    ];

    static getAttributeTypeMap() {
        return GetResponseModel.attributeTypeMap;
    }
}