// src/model/getResultItemModel.ts
/**
 * CyborgDB Service
 * REST API for CyborgDB: The Confidential Vector Database
 *
 * The version of the OpenAPI document: 0.9.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

/**
* Represents an individual item retrieved from the encrypted index.  Attributes:     id (str): The unique identifier of the item.     metadata (Optional[Dict[str, Any]]): Additional metadata associated with the item.     contents (Optional[bytes]): The raw byte contents of the item.     vector (Optional[List[float]]): The vector representation of the item.
*/
export class GetResultItemModel {
    'id': string;
    'metadata'?: object | null;
    'contents'?: RequestFile | null;
    'vector'?: Array<number> | null;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "object"
        },
        {
            "name": "contents",
            "baseName": "contents",
            "type": "RequestFile"
        },
        {
            "name": "vector",
            "baseName": "vector",
            "type": "Array<number>"
        }    ];

    static getAttributeTypeMap() {
        return GetResultItemModel.attributeTypeMap;
    }
}